minimum_code/requirements.txt
matplotlib==3.7.2
numpy==1.25.2
open3d==0.14.1
opencv_python==4.8.0.76
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.11.4
six==1.16.0
smplx==0.1.28
torch==1.12.0
torchgeometry==0.1.2
torchvision==0.13.0
tqdm==4.66.1
